backend/requirements.txt
# Backend FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Base de données MongoDB
motor==3.3.2
pymongo==4.6.0

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0

# Parsing de fichiers
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2
pandas==2.1.4

# Utilitaires
python-dateutil==2.8.2